+++ libs/types/src/lib.rs
@@ -1,4 +1,10 @@
 //! Candid types used by the candid interface of the SOL RPC canister.
+//!
+//! ⚠️ **Build Requirements**
+//!
+//! If you are using the `sol_rpc_types` crate inside a canister, make sure to follow the steps
+//! outlined [here](https://github.com/dfinity/sol-rpc-canister/blob/main/libs/client/README.md#build-requirements)
+//! to ensure your code compiles.
 
 #![forbid(unsafe_code)]
 #![forbid(missing_docs)]