+++ libs/types/src/lifecycle/mod.rs
@@ -2,28 +2,36 @@ use crate::OverrideProvider;
 use candid::{CandidType, Principal};
 use canlog::LogFilter;
 use serde::{Deserialize, Serialize};
+use strum::EnumIter;
 
 /// The installation args for the Solana RPC canister.
 #[derive(Clone, Debug, Default, CandidType, Deserialize)]
 pub struct InstallArgs {
     /// Principals allowed to manage API keys.
+    /// If not specified, only controllers may manage API keys.
     #[serde(rename = "manageApiKeys")]
     pub manage_api_keys: Option<Vec<Principal>>,
     /// Overrides the RPC providers' default URL and HTTP headers.
+    /// If not specified, default URL and HTTP headers are not modified.
     #[serde(rename = "overrideProvider")]
     pub override_provider: Option<OverrideProvider>,
     /// Only log entries matching this filter will be recorded.
+    /// Default is `LogFilter::ShowAll`.
     #[serde(rename = "logFilter")]
     pub log_filter: Option<LogFilter>,
     /// Number of subnet nodes.
+    /// Default is `34` (i.e. the number of nodes in the fiduciary subnet).
     #[serde(rename = "numSubnetNodes")]
     pub num_subnet_nodes: Option<NumSubnetNodes>,
-    /// Mode of operation. Default is `Mode::Normal`.
+    /// Mode of operation.
+    /// Default is `Mode::Normal`.
     pub mode: Option<Mode>,
 }
 
 /// Mode of operation
-#[derive(Debug, Copy, Clone, Default, PartialEq, Eq, CandidType, Deserialize, Serialize)]
+#[derive(
+    Debug, Copy, Clone, Default, PartialEq, Eq, CandidType, Deserialize, EnumIter, Serialize,
+)]
 pub enum Mode {
     #[default]
     /// Normal mode, where cycle payment is required for certain operations.